[mob][photos] Make 'Reassign Me' functional
